// LiveGreenApp/src/app/models/testresult.model.ts
export class TestResult {

    id: number;
    name: string;
    testId: string;
    scenario: string;
    feature: string;
    buildResultKey: string;
    categories: string;
    message: string;
    stackTrace: string;
    success: boolean;
    duration: string;
    log: string;
}